+++ libs/Core/Registry.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace libs\Core;
+
+class Registry
+{
+    private static $data = [];
+
+    public static function set($key, $value)
+    {
+        self::$data[$key] = $value;
+    }
+
+    public static function get($key)
+    {
+        return isset(self::$data[$key]) ? self::$data[$key] : null;
+    }
+}